+++ eras/conway/impl/test/Main.hs
@@ -11,7 +11,6 @@ import qualified Test.Cardano.Ledger.Conway.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Conway.Binary.Regression as Regression
 import qualified Test.Cardano.Ledger.Conway.BinarySpec as Binary
 import qualified Test.Cardano.Ledger.Conway.CommitteeRatifySpec as CommitteeRatify
-import qualified Test.Cardano.Ledger.Conway.Conformance as Conformance
 import qualified Test.Cardano.Ledger.Conway.DRepRatifySpec as DRepRatify
 import qualified Test.Cardano.Ledger.Conway.GenesisSpec as Genesis
 import qualified Test.Cardano.Ledger.Conway.GovActionReorderSpec as GovActionReorder
@@ -44,4 +43,3 @@ main =
       describe "Plutus" $ do
         PlutusSpec.spec
       Regression.spec @Conway
-      Conformance.spec