+++ libs/cardano-ledger-conformance/test/Main.hs
@@ -1,0 +1,10 @@
+module Main (main) where
+
+import Test.Cardano.Ledger.Common
+import qualified Test.Cardano.Ledger.Conformance.Spec.Conway as Conway
+
+main :: IO ()
+main =
+  ledgerTestMain $
+    describe "Conformance" $ do
+      Conway.spec